Combined unit tests into a single project.
