+++ Microsoft.Azure.EventHubs.sln
@@ -1,7 +1,7 @@
 ﻿
 Microsoft Visual Studio Solution File, Format Version 12.00
 # Visual Studio 15
-VisualStudioVersion = 15.0.26403.3
+VisualStudioVersion = 15.0.26430.6
 MinimumVisualStudioVersion = 10.0.40219.1
 Project("{2150E333-8FDC-42A3-9474-1A3956D46DE8}") = "src", "src", "{08E028C1-29E7-42B5-871F-C911DB93E78A}"
 EndProject
@@ -14,11 +14,9 @@ Project("{2150E333-8FDC-42A3-9474-1A3956D46DE8}") = "test", "test", "{AF49C862-C
 EndProject
 Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "Microsoft.Azure.EventHubs", "src\Microsoft.Azure.EventHubs\Microsoft.Azure.EventHubs.csproj", "{126D946D-CE0F-4F14-9F13-8FD7098B81D8}"
 EndProject
-Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "Microsoft.Azure.EventHubs.UnitTests", "test\Microsoft.Azure.EventHubs.UnitTests\Microsoft.Azure.EventHubs.UnitTests.csproj", "{154F7B4C-B998-4FA0-933F-F34DB0CA9B88}"
+Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "Microsoft.Azure.EventHubs.Tests", "test\Microsoft.Azure.EventHubs.Tests\Microsoft.Azure.EventHubs.Tests.csproj", "{154F7B4C-B998-4FA0-933F-F34DB0CA9B88}"
 EndProject
 Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "Microsoft.Azure.EventHubs.Processor", "src\Microsoft.Azure.EventHubs.Processor\Microsoft.Azure.EventHubs.Processor.csproj", "{8C89967A-4E1F-46B0-8458-81B545C822B2}"
-EndProject
-Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "Microsoft.Azure.EventHubs.Processor.UnitTests", "test\Microsoft.Azure.EventHubs.Processor.UnitTests\Microsoft.Azure.EventHubs.Processor.UnitTests.csproj", "{F7F892F4-4490-4BC9-BB18-F42F2C85E345}"
 EndProject
 Global
 	GlobalSection(SolutionConfigurationPlatforms) = preSolution
@@ -38,10 +36,6 @@ Global
 		{8C89967A-4E1F-46B0-8458-81B545C822B2}.Debug|Any CPU.Build.0 = Debug|Any CPU
 		{8C89967A-4E1F-46B0-8458-81B545C822B2}.Release|Any CPU.ActiveCfg = Release|Any CPU
 		{8C89967A-4E1F-46B0-8458-81B545C822B2}.Release|Any CPU.Build.0 = Release|Any CPU
-		{F7F892F4-4490-4BC9-BB18-F42F2C85E345}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
-		{F7F892F4-4490-4BC9-BB18-F42F2C85E345}.Debug|Any CPU.Build.0 = Debug|Any CPU
-		{F7F892F4-4490-4BC9-BB18-F42F2C85E345}.Release|Any CPU.ActiveCfg = Release|Any CPU
-		{F7F892F4-4490-4BC9-BB18-F42F2C85E345}.Release|Any CPU.Build.0 = Release|Any CPU
 	EndGlobalSection
 	GlobalSection(SolutionProperties) = preSolution
 		HideSolutionNode = FALSE
@@ -50,6 +44,5 @@ Global
 		{126D946D-CE0F-4F14-9F13-8FD7098B81D8} = {08E028C1-29E7-42B5-871F-C911DB93E78A}
 		{154F7B4C-B998-4FA0-933F-F34DB0CA9B88} = {AF49C862-CB78-4110-A275-8111B387805D}
 		{8C89967A-4E1F-46B0-8458-81B545C822B2} = {08E028C1-29E7-42B5-871F-C911DB93E78A}
-		{F7F892F4-4490-4BC9-BB18-F42F2C85E345} = {AF49C862-CB78-4110-A275-8111B387805D}
 	EndGlobalSection
 EndGlobal
