+++ test/Microsoft.Azure.EventHubs.Tests/Properties/AssemblyInfo.cs
@@ -10,7 +10,7 @@ using Xunit;
 // associated with an assembly.
 [assembly: AssemblyConfiguration("")]
 [assembly: AssemblyCompany("")]
-[assembly: AssemblyProduct("Microsoft.Azure.EventHubs.UnitTests")]
+[assembly: AssemblyProduct("Microsoft.Azure.EventHubs.Tests")]
 [assembly: AssemblyTrademark("")]
 
 // Setting ComVisible to false makes the types in this assembly not visible
