+++ test/Microsoft.Azure.EventHubs.Tests/TestConstants.cs
@@ -1,0 +1,12 @@
+﻿// Copyright (c) Microsoft. All rights reserved.
+// Licensed under the MIT license. See LICENSE file in the project root for full license information.
+
+namespace Microsoft.Azure.EventHubs.Tests
+{
+    static class TestConstants
+    {
+        internal const string EventHubsConnectionStringEnvironmentVariableName = "azure-event-hubs-dotnet/connectionstring";
+        internal const string StorageConnectionStringEnvironmentVariableName = "azure-event-hubs-dotnet/storageconnectionstring";
+        internal const string DefultEventHubName = "eventhub1";
+    }
+}
